some changes done

+++ app/src/main/java/com/theforum/data/server/areaopinions.java
@@ -46,10 +46,10 @@ public class areaopinions {
     private String topicName;
 
     @com.google.gson.annotations.SerializedName("latitude")
-    private int longitude;
+    private double longitude;
 
     @com.google.gson.annotations.SerializedName("longitude")
-    private int latitude;
+    private double latitude;
 
 
 
@@ -157,19 +157,19 @@ public class areaopinions {
         this.topicName = topicName;
     }
 
-    public int getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(int longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 
-    public int getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(int latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 }
